Rename project files.
