+++ dags/unibas/common/logic/logic_utility.py
