+++ dags/unibas/test/test_logic_transform_nested.py
@@ -1,5 +1,5 @@
 import unittest
-from unibas.common.logic.utility import transform_nested
+from unibas.common.logic.logic_utility import transform_nested
 
 
 class TestTransformNested(unittest.TestCase):
